+++ src/main/java/com/project/domain/reward/dto/response/RewardRespondResponse.java
@@ -1,10 +1,12 @@
-package com.project.domain.mission.dto.response;
+package com.project.domain.reward.dto.response;
 
 import java.time.LocalDateTime;
 
-import com.project.domain.mission.model.RewardRespondResult;
+import com.project.domain.mission.dto.response.CustomerSimpleResponse;
+import com.project.domain.mission.dto.response.MissionItemWithStatusResponse;
+import com.project.domain.reward.model.RewardRespondResult;
 
-/** 보상 요청 응답 처리 결과 DTO다. */
+/** 보상 요청 응답 처리 결과 DTO */
 public record RewardRespondResponse(
         Long requestId,
         String status,
