+++ src/main/java/com/project/domain/reward/repository/RewardRepository.java
@@ -1,7 +1,7 @@
-package com.project.domain.mission.repository;
+package com.project.domain.reward.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.project.domain.mission.entity.Reward;
+import com.project.domain.reward.entity.Reward;
 
 public interface RewardRepository extends JpaRepository<Reward, Long> {}
